+++ src/details/ArborX_DetailsBatchedQueries.hpp
@@ -48,7 +48,7 @@ public:
   // the end.  We decided to keep reversePermutation around for now.
 
   template <typename ExecutionSpace, typename Predicates>
-  static Kokkos::View<size_t *, DeviceType>
+  static Kokkos::View<unsigned int *, DeviceType>
   sortQueriesAlongZOrderCurve(ExecutionSpace const &space,
                               Box const &scene_bounding_box,
                               Predicates const &predicates)
@@ -60,7 +60,7 @@ public:
   }
 
   template <typename ExecutionSpace, typename Predicates>
-  static Kokkos::View<size_t *, DeviceType>
+  static Kokkos::View<unsigned int *, DeviceType>
   sortQueriesAlongZOrderCurve(ExecutionSpace const &space,
                               Kokkos::View<Box const, DeviceType> bounds,
                               Predicates const &predicates)
@@ -87,9 +87,10 @@ public:
   // error: The enclosing parent function ("applyPermutation") for an extended
   // __host__ __device__ lambda must not have deduced return type
   template <typename ExecutionSpace, typename Predicates>
-  static auto applyPermutation(ExecutionSpace const &space,
-                               Kokkos::View<size_t const *, DeviceType> permute,
-                               Predicates const &v)
+  static auto
+  applyPermutation(ExecutionSpace const &space,
+                   Kokkos::View<unsigned int const *, DeviceType> permute,
+                   Predicates const &v)
       -> Kokkos::View<
           std::decay_t<
               decltype(Traits::Access<Predicates, Traits::PredicatesTag>::get(
